# packages/rest/src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"ThreadId": {
				"type": "string",
				"example": "ea4e5e69-fc04-465a-90d2-9f8ff78aa71d"
			},
...
			"ProofRole": {
				"enum": [
					"verifier",
					"prover"
				],
				"type": "string"
			},
...
			"DidCommProofsExchangeRecord": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/RecordId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"type": "string"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					},
					"threadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"parentThreadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"state": {
						"$ref": "#/components/schemas/ProofState"
					},
					"role": {
						"$ref": "#/components/schemas/ProofRole"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"errorMessage": {
						"type": "string"
					},
					"protocolVersion": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"type",
					"threadId",
					"state",
					"role",
					"protocolVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProofProtocolVersion": {
				"type": "string",
				"enum": [
					"v1",
					"v2"
				]
			},
...
			"DidCommProofsAcceptProposalOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"protocolVersion": {
						"$ref": "#/components/schemas/ProofProtocolVersion"
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AcceptAnonCredsProposalOptions"
							},
							"indy": {
								"$ref": "#/components/schemas/AcceptAnonCredsProposalOptions"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"protocolVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsAcceptRequestOptions": {
				"properties": {
					"goalCode": {
						"type": "string"
					},
					"autoAcceptProof": {
						"$ref": "#/components/schemas/AutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"useReturnRoute": {
						"type": "boolean",
						"description": "whether to enable return routing on the send presentation message. This value only\nhas an effect for connectionless exchanges."
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CredentialState": {
				"description": "Issue Credential states as defined in RFC 0036 and RFC 0453",
				"enum": [
					"proposal-sent",
					"proposal-received",
					"offer-sent",
					"offer-received",
					"declined",
					"request-sent",
					"request-received",
					"credential-issued",
					"credential-received",
					"done",
					"abandoned"
				],
				"type": "string"
			},
			"CredentialRole": {
				"enum": [
					"issuer",
					"holder"
				],
				"type": "string"
			},
...
			"DidCommCredentialsExchangeRecord": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/RecordId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"type": "string"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					},
					"threadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"parentThreadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"state": {
						"$ref": "#/components/schemas/CredentialState"
					},
					"role": {
						"$ref": "#/components/schemas/CredentialRole"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/AutoAcceptCredential"
					},
					"revocationNotification": {
						"properties": {
							"comment": {
								"type": "string"
							},
							"revocationDate": {
								"type": "string",
								"format": "date-time"
							}
						},
						"required": [
							"revocationDate"
						],
						"type": "object"
					},
					"errorMessage": {
						"type": "string"
					},
					"protocolVersion": {
						"type": "string"
					},
					"credentials": {
						"items": {
							"$ref": "#/components/schemas/CredentialRecordBinding"
						},
						"type": "array"
					},
					"credentialAttributes": {
						"items": {
							"$ref": "#/components/schemas/CredentialPreviewAttributeOptions"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"type",
					"threadId",
					"state",
					"role",
					"protocolVersion",
					"credentials"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProposeCredentialOptions": {
				"properties": {
					"protocolVersion": {
						"$ref": "#/components/schemas/CredentialProtocolVersion"
					},
					"credentialFormats": {
						"properties": {
							"indy": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/AnonCredsProposeCredentialFormat"
									},
									{
										"$ref": "#/components/schemas/LegacyIndyProposeCredentialFormat"
									}
								]
							},
							"anoncreds": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/AnonCredsProposeCredentialFormat"
									},
									{
										"$ref": "#/components/schemas/LegacyIndyProposeCredentialFormat"
									}
								]
							}
						},
						"type": "object"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/AutoAcceptCredential"
					},
					"comment": {
						"type": "string"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					}
				},
				"required": [
					"protocolVersion",
					"credentialFormats",
					"connectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AcceptCredentialProposalOptions": {
				"properties": {
					"credentialFormats": {
						"properties": {
							"indy": {
								"$ref": "#/components/schemas/AnonCredsAcceptProposalFormat"
							},
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsAcceptProposalFormat"
							}
						},
						"type": "object"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/AutoAcceptCredential"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"tenants": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@credo-ts/rest",
		"version": "0.9.5",
		"description": "Rest API for using Credo over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/proofs/{proofExchangeId}/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommProofsExchangeRecord"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation proposal as verifier by sending an accept proposal message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsAcceptProposalOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/proofs/{proofExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation request as prover by sending an accept request message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsAcceptRequestOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/credentials": {
			"get": {
				"operationId": "FindCredentialsByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DidCommCredentialsExchangeRecord"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"credentials": [],
												"type": "CredentialRecord",
												"role": "holder",
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"state": "offer-sent",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"credentialAttributes": [],
												"protocolVersion": "v1"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Retrieve all credential exchange records by query",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "parentThreadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CredentialState"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CredentialRole"
						}
					}
				]
			}
		},
		"/didcomm/credentials/{credentialExchangeId}": {
			"get": {
				"operationId": "GetCredentialById",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve credential exchange record by credential record id",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Deletes a credential exchange record in the credential repository.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/credentials/propose-credential": {
			"post": {
				"operationId": "ProposeCredential",
				"responses": {
					"200": {
						"description": "CredentialExchangeRecord",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new credential exchange as holder by sending a propose credential message\nto the connection with a specified connection id.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProposeCredentialOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/{credentialExchangeId}/accept-proposal": {
			"post": {
				"operationId": "AcceptProposal",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential proposal as issuer by sending an accept proposal message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialProposalOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/create-offer": {
			"post": {
				"operationId": "CreateOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCommCredentialsCreateOfferResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"credentialExchange": {
												"credentials": [],
												"type": "CredentialRecord",
												"role": "holder",
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"state": "offer-sent",
												"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
												"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
												"credentialAttributes": [],
												"protocolVersion": "v1"
											},
											"message": {
												"@id": "134b27f0-9366-4811-a36b-50bacfe57e61",
												"@type": "https://didcomm.org/issue-credential/1.0/offer-credential"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new credential exchange as issuer by creating a credential offer\nwithout specifying a connection id",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOfferOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/offer-credential": {
			"post": {
				"operationId": "OfferCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate a new credential exchange as issuer by sending a offer credential message\nto the connection with the specified connection id.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OfferCredentialOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/{credentialExchangeId}/accept-offer": {
			"post": {
				"operationId": "AcceptOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential offer as holder by sending an accept offer message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialOfferOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/{credentialExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential request as issuer by sending an accept request message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialRequestOptions"
							}
						}
					}
				}
			}
		},
		"/didcomm/credentials/{credentialExchangeId}/accept-credential": {
			"post": {
				"operationId": "AcceptCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential as holder by sending an accept credential message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
...
		"/didcomm/connections/{connectionId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"type": "ConnectionRecord",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
											"state": "invitation-sent",
											"role": "responder",
											"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
											"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a connection request as inviter by sending a connection response message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}